+++ src/Utility/Streams/Tests/FileOutputStream_ut.cpp
@@ -1,6 +1,6 @@
 #include <cstdlib>
 
-#include "Testing/UnitTest.h"
+#include "Testing/Unit/UnitTest.h"
 
 #include "Utility/Streams/FileOutputStream.h"
