+++ src/Utility/Tests/Segment_ut.cpp
@@ -1,7 +1,7 @@
 #include <numeric>
 #include <vector>
 
-#include "Testing/UnitTest.h"
+#include "Testing/Unit/UnitTest.h"
 
 #include "Utility/Segment.h"
