+++ test/Testing/CMakeLists.txt
@@ -1,3 +1,3 @@
 cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
 
-add_subdirectory(Testing)
+add_subdirectory(Unit)
